Move remaining methods to the crash manager class
