+++ obs-studio-server/CMakeLists.txt
@@ -222,6 +222,8 @@ add_executable(
 	"${PROJECT_SOURCE_DIR}/source/nodeobs_settings.h"
 	"${PROJECT_SOURCE_DIR}/source/util-memory.cpp"
 	"${PROJECT_SOURCE_DIR}/source/util-memory.h"
+	"${PROJECT_SOURCE_DIR}/source/util-crashmanager.cpp"
+	"${PROJECT_SOURCE_DIR}/source/util-crashmanager.h"	
 )
 
 if(WIN32)
